resonance-caretaker/backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
python-dotenv
cryptography
pydantic
python-multipart